+++ src/main/java/me/burninghandsapp/familyportal/models/User.java
@@ -3,12 +3,14 @@ package me.burninghandsapp.familyportal.models;
 import lombok.Getter;
 import lombok.Setter;
 
+import javax.annotation.Priority;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import javax.persistence.criteria.CriteriaBuilder;
 import java.io.Serializable;
 
 
@@ -31,7 +33,7 @@ public class User implements Serializable {
     @Id
     @Column(name = "user_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "username")
     private String userName;
